chore(hadron-document, hadron-type-checker, compass-crud): Add types for hadron packages and move all editors to hadron-document
