+++ packages/compass-components/.prettierignore
@@ -1,2 +1,3 @@
 lib
 .nyc_output
+coverage
