--- packages/compass-crud/src/components/editor/int32.spec.js
@@ -1,14 +1,0 @@
-import { Element } from 'hadron-document';
-import { Int32 } from 'bson';
-import { Int32Editor } from '../editor';
-
-describe('Int32Editor', () => {
-  describe('#size', () => {
-    const element = new Element('field', new Int32(12), false);
-    const int32Editor = new Int32Editor(element);
-
-    it('returns the number of characters', () => {
-      expect(int32Editor.size()).to.equal(2);
-    });
-  });
-});
