--- packages/compass-crud/src/components/editor/string.spec.js
@@ -1,13 +1,0 @@
-import { Element } from 'hadron-document';
-import { StringEditor } from '../editor';
-
-describe('StringEditor', () => {
-  describe('#size', () => {
-    const element = new Element('name', 'test', false);
-    const stringEditor = new StringEditor(element);
-
-    it('returns the number of characters', () => {
-      expect(stringEditor.size()).to.equal(4);
-    });
-  });
-});
