--- packages/compass-crud/src/components/editor/undefined.spec.js
@@ -1,22 +1,0 @@
-import { Element } from 'hadron-document';
-import { UndefinedEditor } from '../editor';
-
-describe('UndefinedEditor', () => {
-  describe('#size', () => {
-    const element = new Element('field', undefined, false);
-    const undefinedEditor = new UndefinedEditor(element);
-
-    it('returns the number of characters', () => {
-      expect(undefinedEditor.size()).to.equal(9);
-    });
-  });
-
-  describe('#value', () => {
-    const element = new Element('field', undefined, false);
-    const undefinedEditor = new UndefinedEditor(element);
-
-    it('returns the string undefined', () => {
-      expect(undefinedEditor.value()).to.equal('undefined');
-    });
-  });
-});
