+++ packages/hadron-document/src/index.js
@@ -1,2 +1,5 @@
 module.exports = require('./document');
+module.exports.DocumentEvents = module.exports.Events;
 module.exports.Element = require('./element');
+module.exports.ElementEvents = module.exports.Element.Events;
+module.exports.ElementEditor = require('./editor');
