+++ packages/hadron-document/src/utils.js
@@ -1,0 +1,6 @@
+const fieldStringLen = (value) => {
+  const length = String(value).length;
+  return length === 0 ? 1 : length;
+};
+
+module.exports = { fieldStringLen };
