+++ packages/hadron-type-checker/test/type-checker.test.js
@@ -722,7 +722,7 @@ describe('TypeChecker', function() {
           'Null',
           'Object',
           'ObjectId',
-          'BSONRegexp',
+          'BSONRegExp',
           'String',
           'BSONSymbol',
           'BSONMap',
@@ -748,7 +748,7 @@ describe('TypeChecker', function() {
           'Null',
           'Object',
           'ObjectId',
-          'BSONRegexp',
+          'BSONRegExp',
           'String',
           'BSONSymbol',
           'BSONMap',
